requirements.txt
matplotlib
numpy
pandas
pyserial
scikit-learn
seaborn
tensorflow
pillow